Add flattened point sequences to wkio
